Weather repository

+++ src/main/java/socialnet/model/Weather.java
@@ -3,12 +3,13 @@ package socialnet.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 @Data
 @AllArgsConstructor
 public class Weather {
-    private Long openWeatherId;
+    private BigInteger openWeatherId;
     private String city;
     private String clouds;
     private Float temp;
